inference-service/requirements.txt
python-dotenv
mem0ai
rank_bm25
fastapi
uvicorn[standard]
openai
qdrant-client
httpx